+++ json_serializable/test/shared_config.dart
@@ -24,9 +24,11 @@ final generatorConfigNonDefaultJson = Map<String, dynamic>.unmodifiable(
     createPerFieldToJson: true,
     disallowUnrecognizedKeys: true,
     explicitToJson: true,
-    fieldRename: FieldRename.kebab,
+    fieldRename: RenameType.kebab,
     ignoreUnannotated: true,
     includeIfNull: false,
     genericArgumentFactories: true,
+    unionDiscriminator: 'runtimeType',
+    unionRename: RenameType.kebab,
   ).toJson(),
 );